requirements.txt
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0

pyspark==3.5.1

sqlalchemy==2.0.32
psycopg2-binary==2.9.9

python-dotenv==1.0.1
pydantic==2.8.2

scikit-learn==1.5.1
statsmodels==0.14.2
prophet==1.1.5

matplotlib==3.9.1
plotly==5.22.0
streamlit==1.37.1
fpdf2==2.7.9
torch==2.3.1
bcrypt==4.1.3
